Epic 3-2. List item (vacancy card) + VacanciesList (+fixes, iteration 2)
